requirements.txt
requests
setuptools
SQLAlchemy
flask-sqlalchemy
flask-login
flask-wtf
python-dotenv